Remove flavor dimentions
