+++ datacapturing/src/androidTest/java/de/cyface/datacapturing/DataCapturingServiceTest.java
