+++ datacapturing/src/androidTest/java/de/cyface/datacapturing/DataCapturingServiceTestWithoutPermission.java
