+++ datacapturing/src/androidTest/java/de/cyface/datacapturing/MovebisTest.java
