+++ datacapturing/src/androidTest/java/de/cyface/datacapturing/PingPongTest.java
