+++ documentation/migration-guide_5.0.0.adoc
@@ -51,11 +51,8 @@ To use it as a dependency in your app you need to:
 +
 ----
  dependencies {
-     # To use the 'movebis' flavour, use: 'datacapturingMovebis'
      implementation "de.cyface:datacapturing:$cyfaceBackendVersion"
-     # To use the 'movebis' flavour, use: 'synchronizationMovebis'
      implementation "de.cyface:synchronization:$cyfaceBackendVersion"
-     # There is only one 'persistence' flavor
      implementation "de.cyface:persistence:$cyfaceBackendVersion"
  }
 ----
