+++ persistence/build.gradle
@@ -26,7 +26,7 @@
  */
 
 plugins {
-    id 'de.cyface.android-publish' version '1.0.8' // Keep version in sync with other modules!
+    id 'de.cyface.android-publish' version '1.1.0' // Keep version in sync with other modules!
 }
 
 android {
