+++ synchronization/build.gradle
@@ -29,7 +29,7 @@ import org.gradle.api.tasks.testing.logging.TestLogEvent
  */
 
 plugins {
-    id 'de.cyface.android-publish' version '1.0.8' // Keep version in sync with other modules!
+    id 'de.cyface.android-publish' version '1.1.0' // Keep version in sync with other modules!
 }
 
 android {
@@ -41,34 +41,6 @@ android {
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionName rootProject.ext.cyfaceBackendVersion
         testInstrumentationRunner rootProject.ext.testInstrumentationRunner
-    }
-
-    // Specifies the flavor dimensions you want to use. The order in which you
-    // list each dimension determines its priority, from highest to lowest,
-    // when Gradle merges variant sources and configurations. You must assign
-    // each product flavor you configure to one of the flavor dimensions.
-    flavorDimensions "project", "mode"
-
-    productFlavors {
-        // Configurations in the "project" product flavors override those in "mode"
-        // flavors and the defaultConfig block. Gradle determines the priority
-        // between flavor dimensions based on the order in which they appear next
-        // to the flavorDimensions property above--the first dimension has a higher
-        // priority than the second, and so on.
-        movebis {
-            dimension "project"
-        }
-        cyface {
-            dimension "project"
-        }
-        mock {
-            dimension "mode"
-            // Excludes flaky tests that are trying to access a Cyface server.
-            testInstrumentationRunnerArgument "notAnnotation", "androidx.test.filters.FlakyTest"
-        }
-        full {
-            dimension "mode"
-        }
     }
 
     buildTypes {
