+++ synchronization/src/androidTest/java/de/cyface/synchronization/MockedHttpConnection.java
