+++ synchronization/src/androidTest/java/de/cyface/synchronization/SyncAdapterTest.java
