+++ synchronization/src/androidTest/java/de/cyface/synchronization/UploadProgressTest.java
@@ -138,7 +138,7 @@ public class UploadProgressTest {
 
         final SharedPreferences preferences = PreferenceManager.getDefaultSharedPreferences(context);
         final SharedPreferences.Editor editor = preferences.edit();
-        editor.putString(SyncService.SYNC_ENDPOINT_URL_SETTINGS_KEY, TEST_API_URL);
+        editor.putString(SyncAdapter.SYNC_ENDPOINT_URL_SETTINGS_KEY, TEST_API_URL);
         editor.apply();
 
         final TestReceiver receiver = new TestReceiver();
