--- synchronization/src/cyface/res/values/strings.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <!-- Used in the authenticator.xml in the cyface flavor -->
-    <string name="cyface" translatable="false">Cyface</string>
-</resources>
