+++ synchronization/src/main/res/ic_launcher-web.png
