+++ synchronization/src/main/res/mipmap-xxxhdpi/ic_launcher_round.png
