+++ synchronization/src/main/res/values/dimens.xml
