+++ synchronization/src/main/res/values/styles.xml
