+++ synchronization/src/main/res/xml/authenticator.xml
