--- synchronization/src/movebis/res/xml/authenticator.xml
@@ -1,2 +1,0 @@
-<account-authenticator xmlns:android="http://schemas.android.com/apk/res/android"
-     android:accountType="de.cyface" />
